# swagger.yaml
openapi: 3.0.3
info:
  title: JetCV NFT API
  description: |
    API completa per la gestione di NFT tramite Crossmint e integrazione Veriff per autenticazione.

    ## 🎯 Funzionalità principali:
    - **NFT Management**: Mint, batch mint, aggiornamento metadati
    - **Wallet Management**: Creazione wallet Solana sicuri
    - **CV Management**: Validazione e creazione file CV con IPFS
    - **Collection Management**: Informazioni collezioni NFT
    - **Veriff Integration**: Autenticazione e verifica documenti

    ## 🔐 Autenticazione:
    - API Key per Crossmint
    - Chiavi pubbliche/private per Veriff

    ## 🌐 Ambienti:
    - Sviluppo: `http://localhost:4000`
    - Staging: `http://18.102.14.247:4000`
    - Produzione: `https://api.jetcv.com`
  version: 3.0.0
  contact:
    name: JetCV Development Team
    email: support@jetcv.com
    url: https://jetcv.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server (local)
  - url: http://18.102.14.247:4000
    description: External server (staging)
  - url: https://api.jetcv.com
    description: Production server
...
paths:
...
  /api/nft/mint:
    post:
      tags: [NFT]
      summary: Mint di un nuovo NFT
      description: |
        Mint di un nuovo NFT tramite Crossmint con integrazione IPFS per CV.

        ## 🔧 Funzionalità:
        - **Validazione indirizzo Ethereum**: Verifica formato 0x[a-fA-F0-9]{40}
        - **Supporto Polygon**: Formato corretto `polygon:address`
        - **Upload IPFS automatico**: Se fornito jsonCV, viene caricato su IPFS
        - **Metadati dinamici**: Supporto per name, description, animation_url, attributes

        ## 📁 Parametri:
        - `to`: Indirizzo Ethereum valido del destinatario
        - `uri`: URI dell'immagine/metadati NFT
        - `metadata`: Metadati opzionali dell'NFT
        - `jsonCV`: JSON del CV da caricare su IPFS (opzionale)

        ## 🚀 Endpoint:
        - **POST** `/api/nft/mint` → Montato su `/nft/mint` nel router
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, uri]
              properties:
                to:
                  type: string
                  pattern: "^0x[a-fA-F0-9]{40}$"
                  example: "0x198c9B45EcFFb65924D20EAeC07776af6975d8B7"
                  description: "Indirizzo Ethereum del destinatario (formato 0x...)"
                uri:
                  type: string
                  example: "https://example.com/image.png"
                  description: "URI dell'immagine o metadati dell'NFT"
                metadata:
                  type: object
                  properties:
                    name:
                      type: string
                      example: "JetCV NFT #1"
                      description: "Nome dell'NFT"
                    description:
                      type: string
                      example: "NFT unico creato tramite JetCV"
                      description: "Descrizione dell'NFT"
                    animation_url:
                      type: string
                      example: "https://example.com/animation.mp4"
                      description: "URL dell'animazione (opzionale)"
                    attributes:
                      type: array
                      items:
                        type: object
                        properties:
                          trait_type:
                            type: string
                            example: "Rarity"
                            description: "Tipo di attributo"
                          value:
                            type: string
                            example: "Legendary"
                            description: "Valore dell'attributo"
                jsonCV:
                  type: object
                  description: "JSON del CV da creare e caricare su IPFS (opzionale)"
                  example:
                    name: "Mario Rossi"
                    email: "mario.rossi@email.com"
                    skills: ["JavaScript", "Node.js", "React"]
                    experience: "5 anni"
                    education: "Laurea in Informatica"
      responses:
        "200":
          description: NFT mintato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    {
                      type: string,
                      example: "NFT mintato con successo tramite Crossmint",
                    }
                  to:
                    { type: string, description: "Indirizzo del destinatario" }
                  uri:
                    type: string
                    description: "URI finale dell'NFT (IPFS se fornito jsonCV)"
                  metadata:
                    type: object
                    description: "Metadati dell'NFT inviati a Crossmint"
                  collectionId:
                    type: string
                    description: "ID della collezione Crossmint"
                  crossmintId:
                    type: string
                    description: "ID dell'NFT su Crossmint"
                  status:
                    type: string
                    description: "Status dell'NFT (onChain.status o status)"
                  chain:
                    type: string
                    example: "polygon"
                    description: "Blockchain utilizzata"
                  contractAddress:
                    type: string
                    nullable: true
                    description: "Indirizzo del contratto smart contract"
                  actionId:
                    type: string
                    nullable: true
                    description: "ID dell'azione Crossmint"
                  ipfs:
                    type: object
                    description: "Dettagli IPFS se fornito jsonCV"
                    properties:
                      cid:
                        type: string
                        description: "Content Identifier IPFS"
                      ipfsUrl:
                        type: string
                        description: "URL IPFS completo (ipfs://CID)"
                      gatewayUrl:
                        type: string
                        description: "URL gateway Lighthouse IPFS"
                      success:
                        type: boolean
                        description: "Se il caricamento IPFS è riuscito"
                      error:
                        type: string
                        nullable: true
                        description: "Messaggio di errore se fallisce"
                  hasCV:
                    type: boolean
                    description: "Se è stato fornito un jsonCV"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Campi 'to' e 'uri' obbligatori"
                    description: "Errore di validazione"
                  details:
                    type: string
                    example: "Indirizzo 'to' non valido - deve essere formato 0x..."
                    description: "Dettagli dell'errore"
                  ipfs:
                    type: object
                    description: "Dettagli IPFS anche in caso di errore"
        "503":
          description: Crossmint non disponibile
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    {
                      type: string,
                      example: "Crossmint API key non configurata",
                    }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Errore durante il minting tramite Crossmint"
                  details:
                    type: string
                    example: "Verifica i parametri e la configurazione"
                  ipfs:
                    type: object
                    description: "Dettagli IPFS anche in caso di errore"
  /api/nft/mint/batch:
    post:
      tags: [NFT]
      summary: Batch mint di NFT
      description: |
        Mint di più NFT in batch tramite Crossmint.

        ## 🔧 Funzionalità:
        - **Validazione indirizzi**: Verifica formato Ethereum per ogni destinatario
        - **Metadati personalizzati**: Supporto per name, description, animation_url, attributes
        - **Supporto CV**: Possibilità di includere jsonCV per ogni NFT
        - **Processamento parallelo**: Creazione simultanea di più NFT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nfts]
              properties:
                nfts:
                  type: array
                  items:
                    type: object
                    required: [to, uri]
                    properties:
                      to:
                        type: string
                        pattern: "^0x[a-fA-F0-9]{40}$"
                        example: "0x198c9B45EcFFb65924D20EAeC07776af6975d8B7"
                        description: "Indirizzo Ethereum del destinatario (formato 0x...)"
                      uri:
                        type: string
                        example: "https://example.com/image.png"
                        description: "URI dell'immagine o metadati dell'NFT"
                      metadata:
                        type: object
                        description: "Metadati dell'NFT"
                        properties:
                          name: { type: string, example: "NFT Name" }
                          description:
                            { type: string, example: "NFT Description" }
                          animation_url:
                            {
                              type: string,
                              example: "https://example.com/animation.mp4",
                            }
                          attributes:
                            type: array
                            items:
                              type: object
                              properties:
                                trait_type: { type: string, example: "Rarity" }
                                value: { type: string, example: "Legendary" }
                      jsonCV:
                        type: object
                        description: "JSON del CV da caricare su IPFS (opzionale)"
                        example:
                          name: "Test User"
                          email: "test@example.com"
                          skills: ["JavaScript", "React"]
                  description: "Array di NFT da mintare"
      responses:
        "200":
          description: Batch mint avviato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    type: string
                    example: "NFT batch creati con successo"
                  batch:
                    type: object
                    description: "Risposta completa da Crossmint"
                    properties:
                      nfts:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string, description: "ID dell'NFT" }
                            status:
                              { type: string, description: "Status dell'NFT" }
                            recipient:
                              { type: string, description: "Destinatario" }
                            metadata:
                              { type: object, description: "Metadati dell'NFT" }
                      total:
                        { type: integer, description: "Numero totale di NFT" }
                      batchId: { type: string, description: "ID del batch" }
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    { type: string, example: "Parametri mancanti o non validi" }
        "503":
          description: Crossmint non disponibile
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    {
                      type: string,
                      example: "Crossmint API key non configurata",
                    }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: "Errore interno del server" }
  /api/nft/status/{crossmintId}:
    get:
      tags: [NFT]
      summary: Stato di un NFT
      description: |
        Recupera lo stato di un NFT tramite Crossmint ID.

        ## 🔧 Funzionalità:
        - **Stato completo**: Status corrente e on-chain dell'NFT
        - **Dati metadati**: Informazioni complete sui metadati
        - **Dettagli blockchain**: Chain, contratto, hash transazione
        - **Informazioni temporali**: Date di creazione e aggiornamento
      parameters:
        - name: crossmintId
          in: path
          required: true
          schema:
            type: string
          description: ID Crossmint dell'NFT
      responses:
        "200":
          description: Stato NFT recuperato
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  nft:
                    type: object
                    description: "Dati completi dell'NFT da Crossmint"
                    properties:
                      id: { type: string, description: "ID dell'NFT" }
                      crossmintId: { type: string, description: "ID Crossmint" }
                      status: { type: string, description: "Status corrente" }
                      metadata:
                        { type: object, description: "Metadati dell'NFT" }
                      recipient: { type: string, description: "Destinatario" }
                      collectionId:
                        { type: string, description: "ID collezione" }
                      onChain:
                        type: object
                        description: "Dati on-chain dell'NFT"
                        properties:
                          status:
                            { type: string, description: "Status on-chain" }
                          chain: { type: string, description: "Blockchain" }
                          contractAddress:
                            { type: string, description: "Contratto" }
                          txHash:
                            { type: string, description: "Hash transazione" }
                      createdAt: { type: string, format: date-time }
                      updatedAt: { type: string, format: date-time }
        "503":
          description: Crossmint non disponibile
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    {
                      type: string,
                      example: "Crossmint API key non configurata",
                    }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: "Errore interno del server" }
  /api/nft/update/{crossmintId}:
    patch:
      tags: [NFT]
      summary: Aggiorna metadati NFT
      description: |
        Aggiorna i metadati di un NFT esistente tramite Crossmint.

        ## 🔧 Funzionalità:
        - **Aggiornamento flessibile**: Supporto per tutti i campi metadati
        - **Validazione dati**: Controllo integrità metadati
        - **Preservazione esistente**: Mantiene i dati non modificati
        - **Risposta completa**: Dati aggiornati dell'NFT
      parameters:
        - name: crossmintId
          in: path
          required: true
          schema:
            type: string
          description: ID Crossmint dell'NFT da aggiornare
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [metadata]
              properties:
                metadata:
                  type: object
                  description: Metadati da aggiornare
                  properties:
                    name:
                      {
                        type: string,
                        example: "Updated NFT Name",
                        description: "Nome dell'NFT",
                      }
                    description:
                      {
                        type: string,
                        example: "New description",
                        description: "Descrizione dell'NFT",
                      }
                    image:
                      {
                        type: string,
                        example: "https://new-image-url.com/image.png",
                        description: "URL dell'immagine",
                      }
                    animation_url:
                      {
                        type: string,
                        description: "URL dell'animazione (opzionale)",
                      }
                    attributes:
                      type: array
                      description: Attributi dell'NFT
                      items:
                        type: object
                        properties:
                          trait_type: { type: string }
                          value: { type: string }
      responses:
        "200":
          description: NFT aggiornato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    type: string
                    example: "NFT aggiornato con successo"
                  nft:
                    type: object
                    description: "Dati aggiornati dell'NFT da Crossmint"
                    properties:
                      id: { type: string, description: "ID dell'NFT" }
                      crossmintId: { type: string, description: "ID Crossmint" }
                      status: { type: string, description: "Status corrente" }
                      metadata:
                        { type: object, description: "Metadati aggiornati" }
                      recipient: { type: string, description: "Destinatario" }
                      collectionId:
                        { type: string, description: "ID collezione" }
                      onChain:
                        type: object
                        description: "Dati on-chain dell'NFT"
                        properties:
                          status:
                            { type: string, description: "Status on-chain" }
                          chain: { type: string, description: "Blockchain" }
                          contractAddress:
                            { type: string, description: "Contratto" }
                          txHash:
                            { type: string, description: "Hash transazione" }
                      createdAt: { type: string, format: date-time }
                      updatedAt: { type: string, format: date-time }
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    { type: string, example: "Parametri mancanti o non validi" }
        "503":
          description: Crossmint non disponibile
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    {
                      type: string,
                      example: "Crossmint API key non configurata",
                    }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: "Errore interno del server" }
  /api/nft/metadata:
    get:
      tags: [NFT]
      summary: Recupera tutti gli NFT della collezione
      description: |
        Recupera tutti gli NFT con i loro metadati dalla collezione Crossmint.

        ## 🔧 Funzionalità:
        - **Paginazione avanzata**: Supporto per page e perPage
        - **Formattazione dati**: Dati formattati per facile utilizzo
        - **Metadati completi**: Nome, descrizione, immagine, attributi
        - **Informazioni temporali**: Date di creazione e aggiornamento
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Numero di pagina
        - name: perPage
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Numero di NFT per pagina
      responses:
        "200":
          description: Tutti gli NFT della collezione recuperati con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    { type: string, example: "NFT recuperati con successo" }
                  success: { type: boolean, example: true }
                  nfts:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, description: "ID dell'NFT" }
                        crossmintId:
                          { type: string, description: "ID Crossmint" }
                        name: { type: string, description: "Nome dell'NFT" }
                        description:
                          { type: string, description: "Descrizione dell'NFT" }
                        image: { type: string, description: "URL immagine" }
                        animation_url:
                          { type: string, description: "URL animazione" }
                        attributes:
                          type: array
                          items:
                            type: object
                            properties:
                              trait_type:
                                { type: string, description: "Tipo attributo" }
                              value:
                                {
                                  type: string,
                                  description: "Valore attributo",
                                }
                        recipient: { type: string, description: "Destinatario" }
                        status: { type: string, description: "Status dell'NFT" }
                        createdAt: { type: string, format: date-time }
                        updatedAt: { type: string, format: date-time }
                  pagination:
                    type: object
                    properties:
                      page: { type: integer, example: 1 }
                      perPage: { type: integer, example: 10 }
                      total: { type: integer, example: 50 }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: "Errore interno del server" }
...
security:
  - ApiKeyAuth: []